feat: Refactor homepage components and enhance terminal demo functionality
