+++ apps/web/app/components/auth/form.vue
