+++ apps/web/app/components/auth/sign-up.vue
