+++ apps/web/app/components/header/index.vue
@@ -1,8 +1,4 @@
 <script setup lang="ts">
-import ColorModeButton from './ColorModeButton.vue';
-import LocaleSelect from './LocaleSelect.vue';
-import UserSelect from './UserSelect.vue';
-
 const session = authClient.useSession();
 </script>
 
@@ -18,9 +14,9 @@ const session = authClient.useSession();
     </template>
 
     <template #right>
-      <ColorModeButton class="light:hover:bg-neutral-100/10" />
-      <LocaleSelect class="light:hover:bg-neutral-100/10" />
-      <UserSelect v-if="session.data?.user" class="light:hover:bg-neutral-100/10" />
+      <HeaderColorModeToggle class="light:hover:bg-neutral-100/10" />
+      <HeaderLocaleSwitcher class="light:hover:bg-neutral-100/10" />
+      <HeaderUserMenu v-if="session.data?.user" class="light:hover:bg-neutral-100/10" />
     </template>
   </UHeader>
 </template>
