+++ apps/web/app/components/header/locale-switcher.vue
