+++ apps/web/app/components/terminal/index.client.vue
