+++ apps/web/app/components/topic/editor/panel.vue
