+++ apps/web/app/components/topic/editor/sidebar.vue
